64-bit fixes

+++ src/bflib_filelst.c
@@ -108,7 +108,7 @@ int LbDataLoad(struct TbLoadFiles *load_file, LoadFilesGetSizeFunc get_size_fn, 
 #ifdef __DEBUG
     LbJustLog("LbDataLoad: filelength %ld for file \"%s\"\n",slength,fname);
 #endif
-    load_file->SLength = get_size_fn? get_size_fn(slength): slength;
+    load_file->SLength = (get_size_fn) ? (long) get_size_fn(slength): slength;
     if (slength <= 0)
         return -101;
     if (!is_static)
